+++ fuzz/fuzz_targets/roundtrip_legacy.rs
@@ -1,0 +1,4 @@
+#![no_main]
+use libfuzzer_sys::fuzz_target;
+
+fuzz_target!(|case: aescrypt_rs_fuzz::LegacyCase| aescrypt_rs_fuzz::fuzz_roundtrip_legacy(&case));