+++ fuzz/fuzz_targets/roundtrip_v3.rs
@@ -1,0 +1,4 @@
+#![no_main]
+use libfuzzer_sys::fuzz_target;
+
+fuzz_target!(|case: aescrypt_rs_fuzz::V3Case| aescrypt_rs_fuzz::fuzz_roundtrip_v3(&case));